A simplified regtest version that is way more stable than the parallel one. Incidentally we think the main problem with threaded execution was that the working directory is set at the process level and so different threads were racing with each other to set it
